poldata/datasources/requirements.txt
Django>=4.2
djangorestframework>=3.14
celery>=5.3
redis>=5.0
tweepy>=4.14.0
facebook-sdk>=3.1.0
instagram-private-api>=1.6.0
beautifulsoup4>=4.12.0
requests>=2.31
pandas>=2.1
ratelimit>=2.2.1
python-dotenv>=1.0
django-redis>=5.4
django-celery-beat>=2.5
matplotlib>=3.8.0